added VerseCheckContaier
